refactor : login refactor use viewmodel
