+++ app/src/main/java/org/android/go/sopt/presentation/viewmodel/SignInViewModel.kt
@@ -1,0 +1,38 @@
+package org.android.go.sopt.presentation.viewmodel
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import org.android.go.sopt.data.remote.ServicePool.signInService
+import org.android.go.sopt.data.remote.dto.RequestSignInDto
+import org.android.go.sopt.data.remote.dto.ResponseSignInDto
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
+
+class SignInViewModel : ViewModel() {
+    val signInResult: MutableLiveData<ResponseSignInDto> = MutableLiveData()
+
+    fun signIn(id: String, password: String) {
+        signInService.login(
+            RequestSignInDto(
+                id,
+                password
+            )
+        ).enqueue(object : Callback<ResponseSignInDto> {
+            override fun onResponse(
+                call: Call<ResponseSignInDto>,
+                response: Response<ResponseSignInDto>,
+            ) {
+                if (response.isSuccessful) {
+                    signInResult.value = response.body()
+                } else {
+                    // TODO: 에러 처리
+                }
+            }
+
+            override fun onFailure(call: Call<ResponseSignInDto>, t: Throwable) {
+                // TODO: 에러 처리
+            }
+        })
+    }
+}
